requirements.txt
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
black==22.3.0
cfgv==3.3.1
click==8.1.3
coverage==6.3
distlib==0.3.4
Django==4.0.5
django-extensions==3.1.5
django-filter==22.1
djangorestframework==3.13.1
execnet==1.9.0
filelock==3.7.1
flake8==4.0.1
identify==2.5.1
iniconfig==1.1.1
mccabe==0.6.1
mock==4.0.3
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.17.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing_extensions==4.2.0
virtualenv==20.14.1
